+++ CHANGELOG.md
@@ -6,6 +6,9 @@ IMPROVEMENTS:
 * variable: Improve the error message returned when using a variable file with an unsupported file extension [[GH-791](https://github.com/hashicorp/nomad-pack/pull/791)]
 * variable: Add support for `optional()` type constraint modifier in variable definitions [[GH-798](https://github.com/hashicorp/nomad-pack/pull/798)]
 
+BUG FIXES:
+* cli: Fix destroy command incorrectly detecting conflicts when multiple jobs share name prefixes (e.g., "Laas-Service" vs "Laas-Service-Regression") [[GH-795](https://github.com/hashicorp/nomad-pack/pull/795)]
+
 ## 0.4.1 (November 13, 2025)
 
 IMPROVEMENTS: